Fix most non-ASCII characters treated as fullwidth
When the rendering engine decides that a character is full-width, it skips a space, leaving it untouched, and continues in the next space. This looks _really_ bad when the character is not, in fact, full-width. (I don't know if it looks good when the character _is_ full-width.) This patch just fixes the detection of full-width characters in the Basic Multilingual Plane. (That last `cc > 0xffee` is also wrong, but didn't seem worth fixing.)