Fix closed set usage in AStar
A node should only be added to the closed set when it is expanded. Before then, the algorithm could still find a shorter path to it, even with a monotone heuristic.

This can result in the same node appearing multiple times in the priority queue, so when pulling a node off the queue, check if it is already closed and if so, skip it. This is not a correctness issue; it merely avoids redundant effort.