Fix implicit any type for Scheduler.next()/remove()
Scheduler<T> takes a type, but its internal representation of the object
tracker was set to `any` instead of `T`. This resulted in `next()`
returning `any`, which lost typing and required unnecessary casting when
it is called..

This also creates an error, as `next()` can also return null, which is not enforced by the type checker.

Lastly, `remove()` item was `any`, rather than `T`, so you lost the
type checking for calling it.

This fixes `next()` and `remove()` to use the expected `T` so that the
type checker catches all of these issues with all of the schedulers.