Adjusting build to work as a TS project reference
This changes some settings in the tsconfig.json file to make this repo
work with the new TypeScript project reference mechanism. It also adds
some partial type declarations in the Term display backend so that
including projects don't have to pull in the @types/node package.

Two notable changes to the tsconfig.json file:

- exactOptionalPropertyTypes is now set to true, so optional properties
  will no longer automatically have undefined added to their types.
- sourceMap is now set to true, so the rot.js build process will
  create source maps; this allows including projects to view the .ts
  file rather than the .js output when following links or in devtools.