Splitting DisplayOptions interface by backend
With this change, each display backend (layout) can specify its own
Options interface as well as its own option defaults. The Display itself
now only stores the literal options passed to the constructor or
setOptions, and the Display.getOptions method takes an optional
parameter for whether to return the effective options (the default) or
only the caller-specified ones. The Term backend uses the ability to
specify defaults to make the Display default to the width and height of
the stdout terminal.