Allow display backends to specify data format
Following up the previous commit, this allows each display backend to
specify a different type for the three components of display data:
characters, fg style, and bg style. This means that a Display created
for a tile backend (which allows arrays for all three components) will
accept either an array or a bare string as arguments to the draw()
method, while a Display created for a term backend (which has no
character overlay support) only allows bare strings in the signature.